refactor over Watchface2. Handlers separated on extern class
