+++ src/app/Settings.cpp
@@ -34,8 +34,8 @@
 
 #include "../system/Network.hpp"
 #include "LogView.hpp"
-#include "Watchface2.hpp"
-// #include "WatchfaceSquare.hpp"
+#include "WatchfaceHandlers.hpp"
+
 extern void StopBLE();
 extern void StartBLE(bool synched=false);
 
@@ -46,14 +46,12 @@ SettingsApplication::~SettingsApplication() {
     NVS.setInt("WifiEnabled",wifiCheck->switchEnabled,false);
 
     NVS.setInt("NTPEnabled",ntpCheck->switchEnabled,false);
-
-    if ( nullptr != Watchface2Application::ntpTask ) { 
-        Watchface2Application::ntpTask->enabled = ntpCheck->switchEnabled;
-    }
+    if (wfhandler.ntpTask != nullptr)
+        wfhandler.ntpTask->enabled = ntpCheck->switchEnabled;
     NVS.setInt("OWeatherEnabled",openweatherCheck->switchEnabled,false);
-    if ( nullptr != Watchface2Application::weatherTask ) {
-        Watchface2Application::weatherTask->enabled = openweatherCheck->switchEnabled;
-    }
+    if (wfhandler.weatherTask != nullptr)
+        wfhandler.weatherTask->enabled = openweatherCheck->switchEnabled;
+    
     if ( bleCheck->switchEnabled ) { StartBLE(); }
     else { StopBLE(); }
     //delete btnHelp;
