+++ src/system/SystemEvents.cpp
@@ -44,8 +44,8 @@ extern TTGOClass *ttgo; // ttgo library
 
 #include "../app/Shutdown.hpp"
 #include "../app/Provisioning2.hpp"
-// #include "../app/WatchfaceAlwaysOn.hpp"
-#include "../app/WatchfaceSquare.hpp"
+#include "../app/WatchfaceAlwaysOn.hpp"
+// #include "../app/WatchfaceSquare.hpp"
 
 #include <HTTPClient.h>
 
@@ -585,6 +585,7 @@ static void BMAEventDoubleTap(void *handler_args, esp_event_base_t base, int32_t
         lEvLog("BMA423: Event: Double tap: Bring up system\n");
         ScreenWake();
         esp_event_post_to(systemEventloopHandler, SYSTEM_EVENTS, SYSTEM_EVENT_WAKE, nullptr, 0, LUNOKIOT_EVENT_MANDATORY_TIME_TICKS);
+        LaunchApplication(new WatchfaceAlwaysOn());
         // LaunchApplication(new WatchfaceSquare());
         //DoSleep();
     }
